Implemented more change requests and reorganized folders
