+++ Assets/Prefabs/Enemy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 73bdbbc1a13e01a48bdae9a81def1b8d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
