+++ Assets/Prefabs/Game Objects/Health Pickup.prefab.meta
