--- Assets/Scripts/LevelSelection.cs
@@ -1,21 +1,0 @@
-using UnityEngine;
-using UnityEngine.UI;
-
-public class LevelSelection : MonoBehaviour
-{
-    [SerializeField]
-    private Button[] lvlButtons;
-
-    void Start()
-    {
-        int levelAt = PlayerPrefs.GetInt("levelAt", 2);
-
-        for (int i = 0; i < lvlButtons.Length; i++)
-        {
-            if ((i + 2) > levelAt)
-            {
-                lvlButtons[i].interactable = false;
-            }
-        }
-    }
-}
