+++ Assets/Scripts/Object Controllers/HPPickupInteract.cs
