+++ ProjectSettings/EditorBuildSettings.asset
@@ -9,9 +9,6 @@ EditorBuildSettings:
     path: 
     guid: 00000000000000000000000000000000
   - enabled: 1
-    path: Assets/Scenes/StartMenu.unity
-    guid: 25ce09e869240a84081359d05e6cc1b7
-  - enabled: 1
     path: Assets/Scenes/LevelSelect.unity
     guid: 30277d48b45e13d4a8eb2a89feddf8cd
   - enabled: 1
@@ -20,4 +17,7 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/Level2.unity
     guid: 8d05261277cd4584e9882efe0ac35f2d
+  - enabled: 1
+    path: Assets/Scenes/StartMenu.unity
+    guid: 25ce09e869240a84081359d05e6cc1b7
   m_configObjects: {}
